fix repeater error styles
